add native feature to sov-modules-impl
