+++ sov-modules/sov-modules-impl/Cargo.toml
@@ -11,5 +11,11 @@ sov-modules-api = { workspace = true }
 sov-modules-macros = { workspace = true }
 sov-state = { workspace = true }
 sovereign-sdk = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true}
+serde = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true}
+
+
+[features]
+default = ["native"]
+serde = ["dep:serde", "dep:serde_json"]
+native = ["serde"]
