+++ sov-modules/sov-modules-impl/src/example_simple/query.rs
@@ -1,5 +1,6 @@
 use super::ValueAdderModule;
 use sov_modules_api::{QueryError, QueryResponse};
+use sovereign_sdk::serial::{Decode, DecodeBorrowed};
 
 pub enum QueryMessage {
     GetValue,
@@ -12,5 +13,23 @@ impl<C: sov_modules_api::Context> ValueAdderModule<C> {
         let data = serde_json::to_vec(&value).unwrap();
 
         Ok(QueryResponse { response: data })
+    }
+}
+
+// Generated
+impl<'de> DecodeBorrowed<'de> for QueryMessage {
+    type Error = ();
+
+    fn decode_from_slice(_: &'de [u8]) -> Result<Self, Self::Error> {
+        todo!()
+    }
+}
+
+// Generated
+impl Decode for QueryMessage {
+    type Error = ();
+
+    fn decode<R: std::io::Read>(_: &mut R) -> Result<Self, <Self as Decode>::Error> {
+        todo!()
     }
 }
